requirements.txt
aniso8601==8.0.0
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.2.0
awscli==1.16.203
backcall==0.1.0
boto3==1.9.231
botocore==1.12.231
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
decorator==4.4.0
docutils==0.14
Flask==1.1.1
flask-restplus==0.13.0
gevent==1.4.0
graphviz==0.8.4
greenlet==0.4.15
idna==2.8
importlib-metadata==0.23
ipython==7.6.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.1.1
kiwisolver==1.1.0
locustio==0.11.0
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.2
mxnet==1.5.1.post0
numpy==1.17.2
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.2.1
python-dateutil==2.8.0
pytube==9.5.1
pytz==2019.3
PyYAML==5.1
pyzmq==18.1.0
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tqdm==4.36.1
traitlets==4.3.2
urllib3==1.25.3
virtualenv==16.6.2
wcwidth==0.1.7
Werkzeug==0.16.0
youtube-dl==2019.7.30
zipp==0.6.0
